Added support for bookmarks, refactored annotations
Annotations are now a type of bookmark, and can be toggled and
centered individually. Future support for traditional image
"bookmarks" (providing location and description) will work
with this interface (as a bookmark object lacking a marker and label).

The bookmarks menu is currently full-screen. An improvement to
a bookmarks sidebar that keep a portion of the image map visible
is desired.