Improve rotation functionality, upgrade OpenLayers
Rotation now takes place directly on the coordinates of a mouse / touch
event object. Due to browser security restrictions, the actual
clientX / clientY event properties are immutable, so the new
clientXr / clientYr properties now contain rotation-compensated
coordinates. A find/replace in OpenLayers (clientX -> clientXr) is
necessary to support this.

All OpenLayers changes were made to the 'rotating-openlayers' git
repository, based on the latest (at the time) development version
of the official OpenLayers. The code was minifed and placed in
'libs/rotating-openlayers.js'.

Any subsequent changes to OpenLayers should be made and committed to
'rotating-openlayers' repository, then minified and coped to this
location.